requirements.txt
# 需安装 yum install python3-devel libevent-devel mysql-devel gcc
# wget https://github.com/apache/rocketmq-client-cpp/releases/download/2.0.0/rocketmq-client-cpp-2.0.0-centos7.x86_64.rpm
# rpm -ivh rocketmq-client-cpp-2.0.0-centos7.x86_64.rpm
# find / -name librocketmq.so
# cp /usr/local/lib/librocketmq.so /usr/lib64/
Django==3.1.2
django-cors-headers==3.5.0
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.12.1
mysqlclient==2.0.1
requests==2.24.0
rocketmq-client-python==2.0.0
sqlparse==0.4.1
urllib3==1.25.10
django-dynamic-db-router==0.3.0
dwebsocket==0.4.2
mysql-connector-python~=8.0.22